add change money user

+++ app/Http/Controllers/OrderController.php
@@ -10,6 +10,7 @@ use App\Http\Services\CartService;
 use App\Http\Services\DetailOrderService;
 use App\Http\Services\OrderService;
 use App\Http\Services\StatusService;
+use App\Http\Services\UserService;
 use App\Models\Status;
 use Dotenv\Util\Str;
 use Illuminate\Http\Request;
@@ -18,18 +19,23 @@ class OrderController extends Controller
 {
     public $data = [];
 
-    public function __construct(OrderService $orderService, DetailOrderService $detailOrderService, CartService $cartService,StatusService $statusService)
+    public function __construct(OrderService $orderService, DetailOrderService $detailOrderService, CartService $cartService,StatusService $statusService, UserService $userService)
     {
         $this->orderService = $orderService;
         $this->detailOrderService = $detailOrderService;
         $this->cartService = $cartService;
         $this->statusService = $statusService;
+        $this->userService = $userService;
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        $this->data['order'] = $this->orderService->getAll();
+        if($request->stt==null)
+            $this->data['order'] = $this->orderService->getAll();
+        else
+            $this->data['order'] = $this->orderService->findByStatusId($request->stt);
         $this->data['count'] = $this->cartService->countProduct();
+        $this->data['statuses'] = $this->statusService->getAll(); 
         return view('user.pages.order.index', $this->data) ;
     }
 
@@ -39,7 +45,10 @@ class OrderController extends Controller
         $data = ['stt_id' => $request->stt_id];
         $this->orderService->update($request->id, $data);
         $this->data['status'] = $this->statusService->getAll(); 
-        $this->data['orders'] = $this->orderService->getAllAdmin($request->stt);
+        if($request->stt==null)
+            $this->data['orders'] = $this->orderService->getAllAdmin();
+        else
+            $this->data['orders'] = $this->orderService->AdminFindByStatusId($request->stt);
         
         return view('admin.pages.order.index', $this->data) ;
     }
@@ -58,23 +67,28 @@ class OrderController extends Controller
      */
     public function create(Request $request)
     {
+        
         $order = new Order();
         $order->user_id = auth()->user()->id;
         $order->stt_id = 1;
         $order->total_price = $request->total_price;
-        $this->orderService->add($order);
-        
-        $data = $this->cartService->getAll();
-        foreach($data as $item){
-            $detailOrder = new DetailOrder();
-            $detailOrder->product_id = $item->product_id;
-            $detailOrder->order_id = $order->id;
-            $detailOrder->quantity = $item->quantity;
-            $detailOrder->current_price	= $item->getProduct->price;
-            $this->detailOrderService->add($detailOrder);
-            $this->cartService->delete($item->id);
+        if($this->orderService->checkMoney($request->total_price)){
+            $this->orderService->add($order);
+            $this->userService->changeMoney($request->total_price);
+            $data = $this->cartService->getAll();
+            foreach($data as $item){
+                $detailOrder = new DetailOrder();
+                $detailOrder->product_id = $item->product_id;
+                $detailOrder->order_id = $order->id;
+                $detailOrder->quantity = $item->quantity;
+                $detailOrder->current_price	= $item->getProduct->price;
+                $this->detailOrderService->add($detailOrder);
+                $this->cartService->delete($item->id);
+            }
+            return redirect(route('user.order.index'));
+        }else{
+            return redirect(route('user.cart.index'));
         }
-        return redirect(route('user.order.index'));
     }
 
     /**
